# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Top One Hundred Billboard Songs",
        "description": "API To Find Top Songs of All Time"
    },
    "host": "localhost:2000",
    "basePath": "/"
    ,
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/toponehundred": {
            "get": {
                "operationId" : "TopOneHundredRouterGet",
                "description" : "Gets all songs or will filter what your looking for",
                "tags": [
                    "Get Songs to Top One Hundred"
                ],
                "summary": "Get a song",  
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TopOneHundred"
                        }
                    }
                },
                "parameters": [
					{
						"description": "The ID of the Song to retrieve",
						"in": "query",
						"name": "topOneHundredId",
						"required": false,
                        "type": "string"
					},
                    {
						"description": "The artist name to retrieve",
						"in": "query",
						"name": "artist",
						"required": false,
                        "type": "string"
					},
                    {
						"description": "The artist type to retrieve",
						"in": "query",
						"name": "artistType",
						"required": false,
                        "type": "string"
					},
                    {
						"description": "The released year to retrieve",
						"in": "query",
						"name": "releaseYear",
						"required": false,
                        "type": "string"
					},
                    {
						"description": "The album to retrieve",
						"in": "query",
						"name": "album",
						"required": false,
                        "type": "string"
					},
                    {
						"description": "Filter by Favorite",
						"in": "query",
						"name": "isFavorite",
						"required": false,
                        "type": "boolean"
					},                   
					{
						"description": "The paging offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"type": "number",
						"format": "double",
						"default": 0
					},
					{
						"description": "The paging limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"type": "number",
						"format": "double",
						"default": 100
					}
				]
            },
            "post": {
                "operationId" : "TopOneHundredRouterPost",
                "description" : "Post a new song",
                "tags": [
                    "Add Song to Top One Hundred"
                ],
                "summary": "Post a song",  
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TopOneHundred"
                        }
                    }
                },
                "parameters": [
					{
						"description": "The Song to create",
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TopOneHundred"
						}
					}
				]
            }
        },
        "/toponehundred/favorite/{id}" : {
            "put": {
                "operationId" : "TopOneHundredRouterPutFavorite",
                "description" : "Update your favorites, this will favorite a song and de-favorite a song, just need to supply the top_one_hundred_id",
                "tags": [
                    "Update 'Favorite' Status on Song"
                ],
                "summary": "Update a favorite on a song",  
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TopOneHundred"
                        }
                    }
                },
                "parameters": [
					{
						"description": "The ID of the Song to retrieve",
						"in": "query",
						"name": "topOneHundredId",
						"required": true,
                        "type": "string"
					}
				]
            }
        }
    },
    "definitions": {
        "TopOneHundred": {
            "properties": {
                "topOneHundredId": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "artistType": {
                    "type": "string"
                },
                "releaseYear": {
                    "type": "string"
                },
                "album": {
                    "type": "string"
                },
                "isFavorite": {
                    "type": "boolean"
                }
            },
            "type": "object",
            "required": [                
            ]
        },
        "PagedResponseBodySongs[]": {
            "description": "",
            "properties": {
                "paging": {
                    "type": "object",
                    "properties": {
                        "rows": {
                            "type": "number",
                            "format": "double",
                            "description": ""
                        },
                        "pages": {
                            "type": "number",
                            "format": "double",
                            "description": ""
                        },
                        "offset": {
                            "type": "number",
                            "format": "double",
                            "description": ""
                        },
                        "limit": {
                            "type": "number",
                            "format": "double",
                            "description": ""
                        },
                        "links": {
                            "type": "object",
                            "properties": {
                                "first": {
                                    "type": "string",
                                    "description": ""
                                },
                                "last": {
                                    "type": "string",
                                    "description": ""
                                },
                                "previous": {
                                    "type": "string",
                                    "description": ""
                                },
                                "next": {
                                    "type": "string",
                                    "description": ""
                                }
                            },
                            "description": ""
                        }
                    },
                    "description": ""
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TopOneHundred"
                    },
                    "description": ""
                }
            },
            "type": "object",
            "required": [
                "paging",
                "data"
            ]
        }
    }
}